pyproject.toml
[project]
name = "palm-mind-trask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.52.0",
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.9.1",
    "nest-asyncio>=1.6.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
]
